pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
name = "avilla-workspace"
dependencies = [
    "graia-broadcast>=0.23.1",
    "graia-amnesia>=0.9.0",
    "creart>=0.3.0",
    "graia-saya>=0.0.18",
    "launart>=0.8.2",
    "loguru>=0.6.0",
    "aiohttp>=3.8.1",
    "nonebot2>=2.0.1",
    "wsproto>=1.2.0",
    "starlette>=0.34.0",
    "uvicorn>=0.25.0",
    "nonechat<1.0.0,>=0.2.0",
    "arclet-alconna>=1.8.5",
    "arclet-alconna-tools>=0.7.0",
    "pygtrie>=2.5.0",
    "selectolax>=0.3.16",
    "satori-python>=0.11.0",
    "lagrange-python>=0.1.1",
]
version = "0"
requires-python = ">=3.9"

[build-system]
requires = ["mina-build<0.6.0"]
build-backend = "mina.backend"

[tool.pdm.build]
includes = ["avilla/"]

[tool.mina]
enabled = true

[tool.black]
include = '\.pyi?$'
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.2.0",
    "mina-build>=0.4.0,<0.6.0",
    "richuru>=0.1.1",
    "ruff>=0.3.2",
    "uvicorn>=0.27.0",
    "devtools>=0.11",
    "ipykernel>=6.23.2",
    "pip>=23.2.1",
    "pdm-mina>=0.3.1",
    "isort>=5.13.2",
]
